Initial release of addon
This addon allows you to announce to a set chat channel when you level
up. It can also be set to announce when you are 25%, 50%, and 75%
completed of a level. It can also announce % to next level at same set
increments.

Options are accessible through interface options frame. You can turn
addon on or off, Select which chat channel to announce to, turn on
percentage reporting, or change percentage reporting to amount left on a
level.

The chat channels currently set up for reporting are:

Guild
Party
Instance
Say
Yell
Local